refactor sprite manager
