+++ src/client/luafunctions.cpp
@@ -38,7 +38,6 @@
 #include "outfit.h"
 #include "player.h"
 #include "protocolgame.h"
-#include "spriteappearances.h"
 #include "spritemanager.h"
 #include "statictext.h"
 #include "thingtypemanager.h"
@@ -119,20 +118,20 @@ void Client::registerLuaFunctions()
 #endif
 
     g_lua.registerSingletonClass("g_sprites");
-    g_lua.bindSingletonFunction("g_sprites", "loadSpr", &SpriteManager::loadSpr, &g_sprites);
+    g_lua.bindSingletonFunction("g_sprites", "loadSpr", &LegacySpriteManager::loadSpr, &g_sprites);
 
-    g_lua.bindSingletonFunction("g_sprites", "unload", &SpriteManager::unload, &g_sprites);
-    g_lua.bindSingletonFunction("g_sprites", "isLoaded", &SpriteManager::isLoaded, &g_sprites);
-    g_lua.bindSingletonFunction("g_sprites", "getSprSignature", &SpriteManager::getSignature, &g_sprites);
-    g_lua.bindSingletonFunction("g_sprites", "getSpritesCount", &SpriteManager::getSpritesCount, &g_sprites);
+    g_lua.bindSingletonFunction("g_sprites", "unload", &LegacySpriteManager::unload, &g_sprites);
+    g_lua.bindSingletonFunction("g_sprites", "isLoaded", &LegacySpriteManager::isLoaded, &g_sprites);
+    g_lua.bindSingletonFunction("g_sprites", "getSprSignature", &LegacySpriteManager::getSignature, &g_sprites);
+    g_lua.bindSingletonFunction("g_sprites", "getSpritesCount", &LegacySpriteManager::getSpritesCount, &g_sprites);
 
 #ifdef FRAMEWORK_EDITOR
-    g_lua.bindSingletonFunction("g_sprites", "saveSpr", &SpriteManager::saveSpr, &g_sprites);
+    g_lua.bindSingletonFunction("g_sprites", "saveSpr", &LegacySpriteManager::saveSpr, &g_sprites);
 #endif
 
     g_lua.registerSingletonClass("g_spriteAppearances");
-    g_lua.bindSingletonFunction("g_spriteAppearances", "saveSpriteToFile", &SpriteAppearances::saveSpriteToFile, &g_spriteAppearances);
-    g_lua.bindSingletonFunction("g_spriteAppearances", "saveSheetToFileBySprite", &SpriteAppearances::saveSheetToFileBySprite, &g_spriteAppearances);
+    g_lua.bindSingletonFunction("g_spriteAppearances", "saveSpriteToFile", &ProtobufSpriteManager::saveSpriteToFile, &g_spriteAppearances);
+    g_lua.bindSingletonFunction("g_spriteAppearances", "saveSheetToFileBySprite", &ProtobufSpriteManager::saveSheetToFileBySprite, &g_spriteAppearances);
 
     g_lua.registerSingletonClass("g_map");
     g_lua.bindSingletonFunction("g_map", "isLookPossible", &Map::isLookPossible, &g_map);
