+++ src/client/thingtype.cpp
@@ -26,7 +26,6 @@
 #include "game.h"
 #include "gameconfig.h"
 #include "lightview.h"
-#include "spriteappearances.h"
 #include "spritemanager.h"
 #include "framework/core/asyncdispatcher.h"
 #include "framework/core/filestream.h"
